some compatibility fixes for xcode 4.x to avoid warnings

+++ vendorsrc/andymatuschak/Sparkle/SUHost.m
@@ -215,7 +215,7 @@
 	OSErr err3 = Gestalt(gestaltSystemVersionBugFix, &bugfix);
 	if (!err1 && !err2 && !err3)
 	{
-		verStr = [NSString stringWithFormat:@"%d.%d.%d", major, minor, bugfix];
+		verStr = [NSString stringWithFormat:@"%d.%d.%d", (int)major, (int)minor, (int)bugfix];
 	}
 	else
 #endif
